requirements.txt
rarfile
matplotlib
matplotlib
pyguitarpro
music21
tqdm